libooni.cpp not needed anymore
