+++ src/ooni/http_invalid_request_line.cpp
@@ -114,5 +114,13 @@ void http_invalid_request_line(Settings options,
         handle_response, options, reactor, logger);
 }
 
+Var<NetTest> HttpInvalidRequestLine::create_test_() {
+    HttpInvalidRequestLine *test = new HttpInvalidRequestLine(options);
+    test->logger = logger;
+    test->reactor = reactor;
+    test->output_filepath = output_filepath;
+    return Var<NetTest>(test);
+}
+
 } // namespace ooni
 } // namespace mk
